Fix model automapper
Some declaratively mapped models were causing the auto mapper to fail.
